+++ src/process.cpp
@@ -19,12 +19,12 @@ void SnapshotProcessor::load_snapshot(const std::string file_name){
   // Now convert this data to the correct format
   int N = strtol(text[0].c_str(), NULL, 10); // # of particles
   // columns are lattice vecs
-  sscanf(text[1].c_str(),"%f %f %f %f %f %f %f %f %f",&a1[0],&a2[0],&a3[0],&a1[1],&a2[1],&a3[1],&a1[2],&a2[2],&a3[2]); // row-major
+  sscanf(text[1].c_str(),"%lf %lf %lf %lf %lf %lf %lf %lf %lf",&a1[0],&a2[0],&a3[0],&a1[1],&a2[1],&a3[1],&a1[2],&a2[2],&a3[2]); // row-major
   positions.reserve(N);
   for(int i = 2; i < N+2; i++){
     // Set the particle's position and orientation from the file
-    std::vector<float> pos(3);
-    sscanf(text[i].c_str(),"%f %f %f",&pos[0],&pos[1],&pos[2]);
+    std::vector<double> pos(3);
+    sscanf(text[i].c_str(),"%lf %lf %lf",&pos[0],&pos[1],&pos[2]);
     positions.push_back(pos);
   }
 }
@@ -35,10 +35,10 @@ void SnapshotProcessor::calculate_order_parameters(size_t max_l){
   // Calculate the bond order parameters
   MSM::MinkowskiStructureCalculator msm;
   msm.load_configuration(positions, a1, a2, a3);
-  q = std::vector<std::vector<float>>(max_l+1, std::vector<float>(N));
-  w = std::vector<std::vector<float>>(max_l+1, std::vector<float>(N));
-  q_av = std::vector<std::vector<float>>(max_l+1, std::vector<float>(N));
-  w_av = std::vector<std::vector<float>>(max_l+1, std::vector<float>(N));
+  q = std::vector<std::vector<double>>(max_l+1, std::vector<double>(N));
+  w = std::vector<std::vector<double>>(max_l+1, std::vector<double>(N));
+  q_av = std::vector<std::vector<double>>(max_l+1, std::vector<double>(N));
+  w_av = std::vector<std::vector<double>>(max_l+1, std::vector<double>(N));
   for(size_t l = 0; l <= max_l; l++){
     q[l] = msm.ql_all(l);
     w[l] = msm.wl_all(l);
@@ -47,13 +47,13 @@ void SnapshotProcessor::calculate_order_parameters(size_t max_l){
   }
 
   // test
-  // float dot = msm.bond_crystallinity(0, 1, 6);
+  // double dot = msm.bond_crystallinity(0, 1, 6);
   // std::cout << dot << '\n';
   // std::vector<unsigned int> all_l(max_l-1); // leave out l=0 and l=1
   // for(unsigned int i = 0, l = 2; l < max_l; i++, l++){
   //   all_l[i] = l;
   // }
-  // float dot_av = msm.bond_crystallinity_lav(0, 1, all_l);
+  // double dot_av = msm.bond_crystallinity_lav(0, 1, all_l);
   // std::cout << dot_av << '\n';
 }
 
@@ -67,7 +67,7 @@ bool SnapshotProcessor::file_exists(const std::string& name)const{
 void SnapshotProcessor::save_boops(
   const std::string target_dir,
   const std::string file_name,
-  std::vector<std::vector<float>>& boop_vector
+  std::vector<std::vector<double>>& boop_vector
 )const{
   size_t N = positions.size();
   size_t max_l = boop_vector.size()-1;
